Style: 상품 상세 교통편 컴포넌트 분리
